use a tiled layout of smaller qr codes
My laptop's webcam seems to have trouble reading the
full-page QR codes. Switching to smaller ones a similar
amount of space, but lets me zoom (or hold it closer to the
camera) to get a better read.

This makes the reading process a bit more tedious,
especially as you have to hit each of the items in order. It
would probably help to cut up the paper so each code could
be showed individually.